requirements.txt
appdirs==1.4.3
autopep8==1.3
Django==1.10.6
django-seed==0.1.7
django-tagulous==0.12.0
djangorestframework==3.6.2
Faker==0.7.10
flake8==3.3.0
flake8-docstrings==1.0.3
flake8-polyfill==1.0.1
importmagic==0.1.7
isort==4.2.5
jedi==0.10.0
mccabe==0.6.1
packaging==16.8
pycodestyle==2.3.1
pydocstyle==1.1.1
pyflakes==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
selenium==3.3.0
six==1.10.0
testfixtures==4.13.5
